refactor: reorganize package layout
