--- src/fundamentus_scrapper/__init__.py
@@ -1,2 +1,0 @@
-from .filters import apply_filters_df, interactive_filter
-
