+++ src/scrapper/scrapper.py
@@ -6,7 +6,7 @@ from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from webdriver_manager.chrome import ChromeDriverManager
 import time
-import src.config
+import config
 
 def get_driver(headless=True):
     options = Options()
